Refactor AlertDialog into its own component
Moved the AlertDialog functionality from ContentDetailsActions.tsx into its own component, DeleteAlertDialog.tsx. This change enhances modularity and readability of the code by breaking down complex components into smaller, more manageable units.
